Services

+++ src/profiler/rebal_list.cpp
@@ -138,7 +138,7 @@ RebalanceCompleteStatistics RebalanceList::statistics(){
             int64_t window_length = profiles[index_start].m_window_length;
             int64_t index_end = index_start; // excl
             RebalanceWindowStatistics window { window_length };
-            while(index_end < (int64_t) profiles.size() && profiles[index_end].m_window_length == window_length && profiles[index_end].m_type == type){
+            while(index_end < (int64_t) profiles.size() && profiles[index_end].m_window_length == (uint64_t) window_length && profiles[index_end].m_type == type){
                 result.m_count++;
                 window.m_count++;
